+++ chainlit_image/foundational-llm-chat_app/config.py
@@ -20,4 +20,7 @@ if os.environ.get('SYSTEM_PROMPT_LIST'):
 MAX_CHARACHERS = None if (not os.getenv("MAX_CHARACHERS") or os.getenv("MAX_CHARACHERS") == "None") else int(os.getenv("MAX_CHARACHERS"))
 MAX_CONTENT_SIZE_MB = None if (not os.getenv("MAX_CONTENT_SIZE_MB") or os.getenv("MAX_CONTENT_SIZE_MB") == "None") else float(os.getenv("MAX_CONTENT_SIZE_MB"))
 
+DYNAMODB_DATA_LAYER_NAME = None if (not os.getenv("DYNAMODB_DATA_LAYER_NAME") or os.getenv("DYNAMODB_DATA_LAYER_NAME") == "None") else str(os.getenv("DYNAMODB_DATA_LAYER_NAME"))
+S3_DATA_LAYER_NAME = None if (not os.getenv("S3_DATA_LAYER_NAME") or os.getenv("S3_DATA_LAYER_NAME") == "None") else str(os.getenv("S3_DATA_LAYER_NAME"))
+
 bedrock_models = None if not os.getenv("BEDROCK_MODELS") else json.loads(os.getenv("BEDROCK_MODELS"))