+++ lib/parameters/index.ts
@@ -9,6 +9,8 @@ export interface ParametersProps {
   readonly bedrock_models_parameter: BedrockModels;
   readonly prefix: string;
   readonly prompts_manager_list: ModelPrompts;
+  readonly dynamodb_dataLayer_name_parameter: string;
+  readonly s3_dataLayer_name_parameter: string;
 }
 
 export class Parameters extends Construct {
@@ -16,6 +18,8 @@ export class Parameters extends Construct {
   public readonly max_characters_parameter: ssm.StringParameter;
   public readonly max_content_size_mb_parameter: ssm.StringParameter;
   public readonly bedrock_models_parameter: ssm.StringParameter;
+  public readonly dynamodb_dataLayer_name_parameter: ssm.StringParameter;
+  public readonly s3_dataLayer_name_parameter: ssm.StringParameter;
 
   constructor(scope: Construct, id: string, props: ParametersProps) {
     super(scope, id);
@@ -41,9 +45,23 @@ export class Parameters extends Construct {
       tier: ssm.ParameterTier.STANDARD,
     });
 
+    this.dynamodb_dataLayer_name_parameter = new ssm.StringParameter(this, "DynamoDBDataLayerNameParameter", {
+      parameterName: `${props.prefix}dynamodb_dataLayer_name_parameter`,
+      description: "DynamoDB table name for data layer",
+      stringValue: props.dynamodb_dataLayer_name_parameter,
+      tier: ssm.ParameterTier.STANDARD,
+    });
+
+    this.s3_dataLayer_name_parameter = new ssm.StringParameter(this, "S3DataLayerNameParameter", {
+      parameterName: `${props.prefix}s3_dataLayer_name_parameter`, 
+      description: "S3 bucket name for data layer",
+      stringValue: props.s3_dataLayer_name_parameter,
+      tier: ssm.ParameterTier.STANDARD,
+    });
+
     // Remove system_prompt field from bedrock_models_parameter
     const bedrock_models_without_system_prompts = Object.entries(props.bedrock_models_parameter).reduce((acc, [key, value]) => {
-      const { system_prompt, ...rest } = value;
+      const { system_prompt, inference_profile, region, ...rest } = value;
       acc[key] = rest;
       return acc;
     }, {} as BedrockModels);